feat(session): consolidate device actions into appium_mobile_device_control
Replace separate shake, notifications, and lock/unlock tools with one action-based tool (lock, unlock, shake, open_notifications) and platform-specific validation.
