+++ src/tools/README.md
@@ -8,10 +8,7 @@ This directory contains all MCP tools available in MCP Appium.
 
 - `create-session.ts` - Create mobile automation sessions
 - `delete-session.ts` - Clean up sessions
-- `open-notifications.ts` - Open notifications panel (Android only)
-- `shake.ts` - iOS Simulator shake via `mobile: shake` (`appium_mobile_shake`; not Android, not physical iOS)
-- `lock.ts` - Lock device (`appium_mobile_lock`); optionally lock for N seconds (Android & iOS)
-- `lock.ts` - Unlock device (`appium_mobile_unlock`)
+- `device-control.ts` - Device controls in one tool (`appium_mobile_device_control`; `action=lock|unlock|shake|open_notifications`)
 - `select-device.ts` - Discover devices and select one (auto-selects if only one found)
 - `file-transfer.ts` - Push/pull files on device (`appium_mobile_file` with `action=push|pull`)
 - `driver-settings.ts` - Read/update Appium driver session settings (`appium_get_settings`, `appium_update_settings`)
