+++ src/tools/index.ts
@@ -22,9 +22,7 @@ import listSessions from './session/list-sessions.js';
 import selectSession from './session/select-session.js';
 import generateLocators from './test-generation/locators.js';
 import selectDevice from './session/select-device.js';
-import openNotifications from './session/open-notifications.js';
-import shakeDevice from './session/shake.js';
-import { lockDevice, unlockDevice } from './session/lock.js';
+import mobileDeviceControl from './session/device-control.js';
 import geolocation from './session/geolocation.js';
 import deviceInfo from './session/device-info.js';
 import fileTransfer from './session/file-transfer.js';
@@ -130,10 +128,7 @@ export default function registerTools(server: FastMCP): void {
   listSessions(server);
   selectSession(server);
   deleteSession(server);
-  openNotifications(server);
-  shakeDevice(server);
-  lockDevice(server);
-  unlockDevice(server);
+  mobileDeviceControl(server);
   geolocation(server);
   deviceInfo(server);
   fileTransfer(server);
